[Refactor] Move BusinessLogic from bizcontroller to bizservice
